Remove requirement that this be used with `npm` 7.x.
This is causing problems for those who still need to use npm 6.

This package doesn't have npm requirements other than the version that is
expected to build its package-lock.json.  This is cruft left over from
copying a package.json from a mononrepo which was meant to ensure people use
the correct version of npm to generate and update the `package-lock.json`
(including the ability to inform Renovate bot what version to use) but I
think we're largely through the npm 6 to 7 migration at this point.